extract griefingCollateral calculations to helpers.ts
